@@ -1,24 +1,25 @@
 import json
 import os
-from typing import Any, Dict, Optional, Type
 from pathlib import Path
+from typing import Any
+
 import msgspec
 from dotenv import load_dotenv
 
-
-__all__ = ["SettingsConfigDict", "BaseSettings"]
+__all__ = ["BaseSettings", "SettingsConfigDict"]
 
 
 class SettingsConfigDict(msgspec.Struct):
-    env_file: Optional[str] = None
+    env_file: str | None = None
     env_file_encoding: str = "utf-8"
     case_sensitive: bool = False
     env_prefix: str = ""
     env_nested_delimiter: str = "__"
 
+
 class BaseSettings:
     model_config: SettingsConfigDict = SettingsConfigDict()
-    
+
     def __init__(self, **values: Any):
         self._load_env_files()
         self._fields = self._get_fields_info()
@@ -33,21 +34,20 @@ class BaseSettings:
             env_path = Path(cls.model_config.env_file)
             if env_path.exists():
                 load_dotenv(
-                    dotenv_path=env_path,
-                    encoding=cls.model_config.env_file_encoding
+                    dotenv_path=env_path, encoding=cls.model_config.env_file_encoding
                 )
 
-    def _get_env_vars(self) -> Dict[str, Any]:
+    def _get_env_vars(self) -> dict[str, Any]:
         """Gets relevant environment variables based on type annotations."""
         env_vars = {}
-        
+
         for field_name, field_type in self.__annotations__.items():
             if field_name == "model_config":
                 continue
-                
+
             env_name = self._get_env_name(field_name)
             env_value = os.environ.get(env_name)
-            
+
             if env_value is not None:
                 # Convert the string value to the appropriate type
                 try:
@@ -55,18 +55,18 @@ class BaseSettings:
                     env_vars[field_name] = converted_value
                 except (ValueError, json.JSONDecodeError) as e:
                     raise ValueError(
-                        f"Error parsing environment variable {env_name}: {str(e)}"
+                        f"Error parsing environment variable {env_name}: {e!s}"
                     )
-        
+
         return env_vars
 
-    def _get_fields_info(self) -> Dict[str, Any]:
+    def _get_fields_info(self) -> dict[str, Any]:
         """Gets information about fields, including Field settings."""
         fields = {}
         for field_name in self.__annotations__:
             if field_name == "model_config":
                 continue
-            
+
             # Checks if there is a default value defined with Field
             field_value = getattr(self.__class__, field_name, None)
             if isinstance(field_value, msgspec.inspect.Field):
@@ -75,9 +75,14 @@ class BaseSettings:
                     "field": field_value,
                     "name": field_value.name or field_name,
                     "has_default": field_value.default is not msgspec.NODEFAULT,
-                    "default": field_value.default if field_value.default is not msgspec.NODEFAULT else None,
-                    "has_default_factory": field_value.default_factory is not msgspec.NODEFAULT,
-                    "default_factory": field_value.default_factory if field_value.default_factory is not msgspec.NODEFAULT else None,
+                    "default": field_value.default
+                    if field_value.default is not msgspec.NODEFAULT
+                    else None,
+                    "has_default_factory": field_value.default_factory
+                    is not msgspec.NODEFAULT,
+                    "default_factory": field_value.default_factory
+                    if field_value.default_factory is not msgspec.NODEFAULT
+                    else None,
                 }
             else:
                 fields[field_name] = {
@@ -85,7 +90,9 @@ class BaseSettings:
                     "field": None,
                     "name": field_name,
                     "has_default": hasattr(self.__class__, field_name),
-                    "default": field_value if hasattr(self.__class__, field_name) else None,
+                    "default": field_value
+                    if hasattr(self.__class__, field_name)
+                    else None,
                     "has_default_factory": False,
                     "default_factory": None,
                 }
@@ -101,13 +108,15 @@ class BaseSettings:
             name = f"{self.model_config.env_prefix}{name}"
         return name
 
-    def _convert_env_value(self, value: str, field_type: Type) -> Any:
+    def _convert_env_value(self, value: str, field_type: type) -> Any:
         """Converts an environment variable string to the appropriate type."""
         if field_type == bool:
             return value.lower() in ("true", "1", "t", "y", "yes")
         elif field_type == int or str(field_type).startswith("typing.Optional[int]"):
             return int(value)
-        elif field_type == float or str(field_type).startswith("typing.Optional[float]"):
+        elif field_type == float or str(field_type).startswith(
+            "typing.Optional[float]"
+        ):
             return float(value)
         elif field_type == list or str(field_type).startswith("typing.List"):
             if value.startswith("[") and value.endswith("]"):
@@ -116,7 +125,9 @@ class BaseSettings:
         elif field_type == dict or str(field_type).startswith("typing.Dict"):
             return msgspec.json.decode(value.encode())
         # For complex types (like msgspec Structs)
-        elif isinstance(msgspec.inspect.type_info(field_type), msgspec.inspect.StructType):
+        elif isinstance(
+            msgspec.inspect.type_info(field_type), msgspec.inspect.StructType
+        ):
             return msgspec.json.decode(value.encode(), type=field_type)
         # For other types, returns the original string
         return value
@@ -130,48 +141,53 @@ class BaseSettings:
             return field_info["default"]
         return None
 
-    def _validate_and_set_values(self, values: Dict[str, Any]):
+    def _validate_and_set_values(self, values: dict[str, Any]):
         """Validate and set values using msgspec."""
         for field_name, field_info in self._fields.items():
             value = values.get(field_name)
-            
+
             if value is None:
                 value = self._get_field_default(field_name)
-            
+
             if value is not None:
                 try:
                     validated_value = msgspec.convert(value, field_info["type"])
                     setattr(self, field_name, validated_value)
                 except msgspec.ValidationError as e:
-                    raise ValueError(
-                        f"Validation error for field {field_name}: {str(e)}"
-                    )
-            elif not field_info["has_default"] and not field_info["has_default_factory"]:
+                    raise ValueError(f"Validation error for field {field_name}: {e!s}")
+            elif (
+                not field_info["has_default"] and not field_info["has_default_factory"]
+            ):
                 raise ValueError(f"Missing required field: {field_name}")
 
         # Stores schema after validation
         self._schema = self._generate_schema()
 
-    def _generate_schema(self) -> Dict[str, Any]:
+    def _generate_schema(self) -> dict[str, Any]:
         """Generates the JSON Schema for the class."""
+
         def schema_hook(typ):
             if typ is self.__class__:
                 return {
                     "type": "object",
                     "properties": {
-                        field_name: msgspec.json.schema(field_info["type"], schema_hook=schema_hook)
+                        field_name: msgspec.json.schema(
+                            field_info["type"], schema_hook=schema_hook
+                        )
                         for field_name, field_info in self._fields.items()
                     },
                     "required": [
-                        field_name for field_name, field_info in self._fields.items()
-                        if not field_info["has_default"] and not field_info["has_default_factory"]
+                        field_name
+                        for field_name, field_info in self._fields.items()
+                        if not field_info["has_default"]
+                        and not field_info["has_default_factory"]
                     ],
                 }
             return None
 
         return msgspec.json.schema(self.__class__, schema_hook=schema_hook)
 
-    def model_dump(self) -> Dict[str, Any]:
+    def model_dump(self) -> dict[str, Any]:
         """Returns data as a dict."""
         return {
             field_name: getattr(self, field_name)
@@ -183,6 +199,6 @@ class BaseSettings:
         """Returns data as a JSON string."""
         return msgspec.json.encode(self.model_dump()).decode()
 
-    def schema(self) -> Dict[str, Any]:
+    def schema(self) -> dict[str, Any]:
         """Returns the JSON schema of the data."""
         return self._schema
