@@ -1,7 +1,6 @@
 from .settings import BaseSettings, SettingsConfigDict
 
-
 __all__ = [
-    "BaseSettings", 
+    "BaseSettings",
     "SettingsConfigDict",
 ]
